// src/types/validate-type.enum.ts
enum ValidateTypeEnum {
  MinLength = 'MinLength',
  Maxlength = 'Maxlength',
  IsDateString = 'IsDateString',
  IsEnum = 'IsEnum',
  IsMongoId = 'IsMongoId',
  ArrayMinSize = 'ArrayMinSize',
  ArrayMaxSize = 'ArrayMaxSize',
  IsBoolean = 'IsBoolean',
  Min = 'Min',
  Max = 'Max',
  ValidateNested = 'ValidateNested',
  IsEmail = 'IsEmail',
  IsJWT = 'IsJWT'
}

export default ValidateTypeEnum;